+++ src/routes/callback/callbackCode.ts
@@ -1,0 +1,5 @@
+export function retrieveCode(){
+    const queryString = window.location.search;
+    const urlParams = new URLSearchParams(queryString);
+    return urlParams.get('code');
+}
