+++ tests/playwright/slider.test.ts
